Split up FilesTests into FileIOTests and FileMiddlewareTests
